Clocks have phase offset, all tick in lockstep

+++ src/com/cburch/logisim/instance/InstanceState.java
@@ -54,7 +54,7 @@ public interface InstanceState {
 
   public Project getProject();
 
-  public long getTickCount();
+  public int getTickCount();
 
   public boolean isCircuitRoot();
